Fix voltage calculation. Implement voltage correction state.

+++ Firmware/Tach/power_monitor.h
@@ -3,6 +3,7 @@
 
 void power_monitor_init();
 uint16_t power_monitor_get_voltage();
-void power_monitor_set_voltage_compensation(uint8_t compensation);
+void power_monitor_set_voltage_compensation(int16_t compensation);
+
 
 #endif
